refactor(web): extract conversation management hook
