+++ web/lib/use-conversations.tsx
@@ -1,9 +1,8 @@
 'use client';
 
 import { useCallback } from 'react';
-import { flushSync } from 'react-dom';
 
-import type { FeedbackType, MyUIMessage } from '@/lib/api-types';
+import type { FeedbackType } from '@/lib/api-types';
 
 import { fireAndForget } from '@/lib/async';
 import { renderAnswerActions } from '@/lib/conversation-actions';
@@ -11,16 +10,10 @@ import {
   applyFeedback,
   previewRegeneration,
 } from '@/lib/conversation-message-state';
-import { deriveTitle } from '@/lib/messages';
-import {
-  clearChatConversations,
-  deleteChatConversation,
-  DeleteChatConversationError,
-  saveChatConversation,
-} from '@/lib/transport';
 import { useUiStore } from '@/lib/ui-store';
 import { useConversationChatRuntime } from '@/lib/use-conversation-chat-runtime';
 import { useConversationList } from '@/lib/use-conversation-list';
+import { useConversationManagement } from '@/lib/use-conversation-management';
 import { useGeneratedTitle } from '@/lib/use-generated-title';
 import { useStopChat } from '@/lib/use-stop-chat';
 
@@ -59,169 +52,25 @@ export const useConversations = (
   const { applyGeneratedTitle, generatingTitleId, handleGenerateTitle } =
     useGeneratedTitle({ conversations, modelRef, refreshConversations });
   const handleStop = useStopChat({ convoIdRef, messages, model, stop });
-
-  const handleNewChat = useCallback(() => {
-    if (status !== 'ready') {
-      fireAndForget(
-        (async () => {
-          await handleStop('local-first');
-          setActiveId(null);
-          setMessages([]);
-          setActiveError(undefined);
-          convoIdRef.current = null;
-        })(),
-      );
-      return;
-    }
-    setActiveId(null);
-    setMessages([]);
-    setActiveError(undefined);
-    convoIdRef.current = null;
-  }, [
+  const {
+    handleClearAll,
+    handleDelete,
+    handleNewChat,
+    handleRename,
+    handleSelect,
+    submitMessage,
+  } = useConversationManagement({
+    applyGeneratedTitle,
     convoIdRef,
     handleStop,
-    setActiveError,
-    setActiveId,
-    setMessages,
-    status,
-  ]);
-
-  const handleSubmit = useCallback(
-    async (text: string) => {
-      setActiveError(undefined);
-      const existing = convoIdRef.current;
-      let cid = existing;
-      let expectedTitle: null | string = null;
-      if (!existing) {
-        expectedTitle = deriveTitle(text);
-        cid = crypto.randomUUID();
-        await saveChatConversation({
-          id: cid,
-          model,
-          title: expectedTitle,
-        });
-        // eslint-disable-next-line require-atomic-updates -- fresh id, not a read-modify-write race
-        convoIdRef.current = cid;
-        // eslint-disable-next-line @eslint-react/dom-no-flush-sync -- conversation id must be committed before useChat resumes the new stream
-        flushSync(() => {
-          setActiveId(cid);
-        });
-        await refreshConversations();
-      }
-      if (!cid) {
-        return;
-      }
-      const userMessage: MyUIMessage = {
-        id: crypto.randomUUID(),
-        metadata: {},
-        parts: [{ text, type: 'text' }],
-        role: 'user',
-      };
-      if (expectedTitle !== null) {
-        fireAndForget(applyGeneratedTitle(cid, [userMessage], expectedTitle));
-      }
-      fireAndForget(sendMessageRef.current(userMessage));
-    },
-    [
-      applyGeneratedTitle,
-      convoIdRef,
-      model,
-      refreshConversations,
-      sendMessageRef,
-      setActiveError,
-      setActiveId,
-    ],
-  );
-
-  const handleSelect = useCallback(
-    (id: string) => {
-      if (id !== convoIdRef.current && status !== 'ready') {
-        fireAndForget(
-          (async () => {
-            await handleStop('local-first');
-            convoIdRef.current = id;
-            setMessages([]);
-            setActiveId(id);
-          })(),
-        );
-        return;
-      }
-      convoIdRef.current = id;
-      setMessages([]);
-      setActiveId(id);
-    },
-    [convoIdRef, handleStop, setActiveId, setMessages, status],
-  );
-
-  const deleteConversationEverywhere = useCallback(
-    async (id: string) => {
-      const deletingActiveConversation = convoIdRef.current === id;
-      if (deletingActiveConversation && status !== 'ready') {
-        await handleStop();
-      }
-
-      try {
-        await deleteChatConversation(id);
-      } catch (error) {
-        const serverConversationAlreadyDeleted =
-          error instanceof DeleteChatConversationError && error.status === 404;
-
-        if (!serverConversationAlreadyDeleted) {
-          throw error;
-        }
-      }
-      if (deletingActiveConversation && convoIdRef.current === id) {
-        setActiveId(null);
-        setMessages([]);
-        setActiveError(undefined);
-        convoIdRef.current = null;
-      }
-      await refreshConversations();
-    },
-    [
-      convoIdRef,
-      handleStop,
-      refreshConversations,
-      setActiveError,
-      setActiveId,
-      setMessages,
-      status,
-    ],
-  );
-
-  const handleDelete = useCallback(
-    (id: string) => {
-      fireAndForget(deleteConversationEverywhere(id));
-    },
-    [deleteConversationEverywhere],
-  );
-
-  const handleClearAll = useCallback(async () => {
-    if (status !== 'ready') {
-      await handleStop();
-    }
-    await clearChatConversations();
-    setActiveId(null);
-    setMessages([]);
-    setActiveError(undefined);
-    convoIdRef.current = null;
-    await refreshConversations();
-  }, [
-    convoIdRef,
-    handleStop,
+    model,
     refreshConversations,
+    sendMessageRef,
     setActiveError,
     setActiveId,
     setMessages,
     status,
-  ]);
-
-  const submitMessage = useCallback(
-    (text: string) => {
-      fireAndForget(handleSubmit(text));
-    },
-    [handleSubmit],
-  );
+  });
 
   const retry = useCallback(() => {
     if (disabled) {
@@ -250,14 +99,6 @@ export const useConversations = (
       setActiveStatus,
       setRegeneratingMessageId,
     ],
-  );
-
-  const handleRename = useCallback(
-    async (id: string, title: string) => {
-      await saveChatConversation({ id, title });
-      await refreshConversations();
-    },
-    [refreshConversations],
   );
 
   const recordFeedback = useCallback(
